Mock (#4)
* Add cmdline support for --mock-dir

* comment out that build for now

Github actions should handle this for us.

* remove this AI garbage.

* Partial collector fixes

* notes for future limits on program output size

* not sure what this is about

I need to improve this for G to reference anyway.

* Use go style error handling rather than looking at the outputs

This convention harkened back to the shell version of clu....

* Update the DI vars

Is it really DI if its a global and not a function argument?

* Update callers to new conventions.

* Create new go versions of mocked input files.

AI: gpt-5 mini, I think it did a decent job.

* Update testdata files to get ready for full program testing.

* Finish out the mock mode - this is tested and works!

* Implement  dependancy injection from main->subcmd run commands to make full e2e testing possible.

This looks weird to me but it works.

* Implement coverage report for our testing code.

* remove old code

* debug the collector and add a basic AI generated test

Not sure what original problem was, but this works.

* Shift the default value to kong itself rather than as explicit code.

* Rename function

---------

Co-authored-by: Hunter Matthews <hunter.matthews@nih.gov>